# src/Juna.SKS.Package.Services/wwwroot/swagger-original.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Parcel Logistics Service",
    "contact": {
      "name": "SKS",
      "url": "http://www.technikum-wien.at/"
    },
    "version": "1.20.2"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/parcel/{trackingId}/reportHop/{code}": {
      "post": {
        "tags": [ "staff" ],
        "summary": "Report that a Parcel has arrived at a certain hop either Warehouse or Truck.\n",
        "operationId": "reportParcelHop",
        "parameters": [
          {
            "name": "trackingId",
            "in": "path",
            "description": "The tracking ID of the parcel.\nE.g. PYJRB4HZ6\n",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "pattern": "^[A-Z0-9]{9}$",
              "type": "string"
            }
          },
          {
            "name": "code",
            "in": "path",
            "description": "The Code of the hop (Warehouse or Truck).",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "pattern": "^[A-Z]{4}\\d{1,4}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully reported hop."
          },
          "400": {
            "description": "The operation failed due to an error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Parcel does not exist with this tracking ID or hop with code not found.\n"
          }
        }
      }
    },
...
    "/parcel/{trackingId}/webhooks": {
      "get": {
        "tags": [ "parcelWebhook" ],
        "summary": "Get all registered subscriptions for the parcel webhook.",
        "operationId": "listParcelWebhooks",
        "parameters": [
          {
            "name": "trackingId",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "pattern": "^[A-Z0-9]{9}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of webooks for the `trackingId`",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/webhookResponses"
                }
              }
            }
          },
          "404": {
            "description": "No parcel found with that tracking ID.",
            "content": {}
          }
        }
      },
      "post": {
        "tags": [ "parcelWebhook" ],
        "summary": "Subscribe to a webhook notification for the specific parcel.",
        "operationId": "subscribeParcelWebhook",
        "parameters": [
          {
            "name": "trackingId",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "pattern": "^[A-Z0-9]{9}$",
              "type": "string"
            }
          },
          {
            "name": "url",
            "in": "query",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/webhookResponse"
                }
              }
            }
          },
          "404": {
            "description": "No parcel found with that tracking ID.",
            "content": {}
          }
        },
        "callbacks": {
          "myevent": {
            "{$request.query.url}?trackingId={$request.path.trackingId}": {
              "post": {
                "requestBody": {
                  "content": {
                    "application/json": {
                      "schema": {
                        "$ref": "#/components/schemas/webhookMessage"
                      }
                    }
                  }
                },
                "responses": {
                  "200": {
                    "description": "OK",
                    "content": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/parcel/webhooks/{id}": {
      "delete": {
        "tags": [ "parcelWebhook" ],
        "summary": "Remove an existing webhook subscription.",
        "operationId": "unsubscribeParcelWebhook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "404": {
            "description": "Subscription does not exist.",
            "content": {}
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "error": {
        "required": [ "errorMessage" ],
        "type": "object",
        "properties": {
          "errorMessage": {
            "type": "string",
            "description": "The error message."
          }
        }
      },
...
      "hopArrival": {
        "required": [ "code", "dateTime", "description" ],
        "type": "object",
        "properties": {
          "code": {
            "pattern": "^[A-Z]{4}\\d{1,4}$",
            "type": "string",
            "description": "Unique CODE of the hop."
          },
          "description": {
            "type": "string",
            "description": "Description of the hop."
          },
          "dateTime": {
            "type": "string",
            "description": "The date/time the parcel arrived at the hop.",
            "format": "date-time"
          }
        },
        "example": {
          "dateTime": "2000-01-23T04:56:07.000+00:00",
          "code": "code",
          "description": "description"
        }
      },
      "trackingInformation": {
        "required": [ "futureHops", "state", "visitedHops" ],
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "description": "State of the parcel.",
            "enum": [ "Pickup", "InTransport", "InTruckDelivery", "Transferred", "Delivered" ]
          },
          "visitedHops": {
            "type": "array",
            "description": "Hops visited in the past.",
            "items": {
              "$ref": "#/components/schemas/hopArrival"
            }
          },
          "futureHops": {
            "type": "array",
            "description": "Hops coming up in the future - their times are estimations.",
            "items": {
              "$ref": "#/components/schemas/hopArrival"
            }
          }
        },
        "example": {
          "visitedHops": [
            {
              "dateTime": "2000-01-23T04:56:07.000+00:00",
              "code": "code",
              "description": "description"
            },
            {
              "dateTime": "2000-01-23T04:56:07.000+00:00",
              "code": "code",
              "description": "description"
            }
          ],
          "futureHops": [ null, null ],
          "state": "Pickup"
        }
      },
...
      "webhookResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "trackingId": {
            "pattern": "^[A-Z0-9]{9}$",
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "created_at": "2000-01-23T04:56:07.000+00:00",
          "id": 0,
          "url": "url",
          "trackingId": "trackingId"
        }
      },
      "webhookResponses": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/webhookResponse"
        }
      },
      "webhookMessage": {
        "allOf": [
          {
            "$ref": "#/components/schemas/trackingInformation"
          },
          {
            "properties": {
              "trackingId": {
                "pattern": "^[A-Z0-9]{9}$",
                "type": "string"
              }
            }
          }
        ]
      },
...
    },
...
  }
}